renamed DeleteVehicle to DeleteResource
