+++ src/components/DeleteResource/DeleteResource.js
